pyproject.toml
[project]
name = "stock-analysis-agent"
version = "0.1.0"
description = "Sophisticated agentic stock analysis system using Google ADK"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=0.1.0",
    "yfinance>=0.2.18",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0"
]
